Fix action-buffer-knn-bug
Fix a bug where the following exception at the begining of the training
sometimes gets thrown:

ValueError: k must be less than or equal to the number of training
points

This can happen when an estimation of Q(s,a) by the knn-algorithm is
perfomed and the number of elements in the action-buffer for the
particular action is less than k.

This fix makes sure that such actions are executed instantly when an
estimation should happen to make sure that every action-buffer has
at least k elements as soon as possible.